+++ src/main/java/usyd/mingyi/animalcare/pojo/Comment.java
@@ -10,13 +10,13 @@ import java.io.Serializable;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Comment implements Serializable {
-    private Integer id;
-    private Integer postId;
+    private Long id;
+    private Long postId;
     private Long commentTime;
     private String commentContent;
     private String userAvatar;
     private String nickName;
     private String userName;
-    private int userId;
+    private Long userId;
 
 }
